[doc] reorganize downstream tasks
