+++ examples/downstream/paper_segmentation/model.py
